Change the PC to Int to avoid conversions
Although a Short is enough to address all of the memory, it causes
constant conversions while accessing the actual memory implementation.

+++ src/chipeitlib/src/main/java/es/chipeit/lib/core/Registers.kt
@@ -5,7 +5,7 @@ import es.chipeit.lib.interfaces.IRegisters
 
 internal class Registers(override val v: IMemory<Byte>) : IRegisters {
     override var i: Short = 0x00
-    override var pc: Short = 0x00
+    override var pc: Int = 0x00
 
     override var sp: Byte = 0x0
